refactor: change module dir
