+++ packages/x6/src/background/main.ts
