+++ packages/x6/src/filter/brightness.ts
