+++ packages/x6/src/filter/gray-scale.ts
