+++ packages/x6/src/filter/invert.ts
