+++ packages/x6/src/filter/outline.ts
