+++ packages/x6/src/grid/double-mesh.ts
@@ -1,4 +1,4 @@
-import { Dom } from '../../util'
+import { Dom } from '../util'
 import { Grid } from './index'
 
 export interface DoubleMeshOptions extends Grid.Options {
