+++ packages/x6/src/grid/fixed-dot.ts
@@ -1,4 +1,4 @@
-import { Dom } from '../../util'
+import { Dom } from '../util'
 import { Grid } from './index'
 
 export interface FixedDotOptions extends Grid.Options {}
