+++ packages/x6/src/highlighter/main.ts
