+++ packages/x6/src/style/index.less
