--- packages/x6/src/style/index.ts
@@ -1,4 +1,0 @@
-export * from './highlighter'
-export * from './background'
-export * from './grid'
-export * from './filter'
